proyecto_seminario2025-main/backend/requirements.txt
Django==5.2.6
djangorestframework==3.14.0
django-cors-headers==4.3.1
mysqlclient==2.2.0
python-dotenv==1.0.0
Pillow==10.1.0
PyJWT==2.8.0